"Feign Client Crap"

+++ src/main/java/com/laman/paydaytradesystem/model/dto/response/DepositResponse.java
@@ -3,6 +3,8 @@ package com.laman.paydaytradesystem.model.dto.response;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -10,5 +12,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class DepositResponse {
 
-    private Double newBalance;
+    private BigDecimal newBalance;
 }
