+++ src/main/java/com/laman/paydaytradesystem/service/DepositService.java
@@ -2,8 +2,10 @@ package com.laman.paydaytradesystem.service;
 
 import com.laman.paydaytradesystem.model.dto.response.DepositResponse;
 
+import java.math.BigDecimal;
+
 public interface DepositService {
 
-    DepositResponse depositCash(Long userId, Double amount);
+    DepositResponse depositCash(Long userId, BigDecimal amount);
 
 }
